+++ src/test/java/com/martak/adventofcode/intcodecomputer/CodeComputerTest.java
@@ -9,7 +9,7 @@ public class CodeComputerTest {
     @Test
     public void test1a() {
         int[] diagnosticTests = new int[]{3,12,6,12,15,1,13,14,13,4,13,99,-1,0,1,9};
-        Context context = new Context(0, diagnosticTests, new int[]{0}, "");
+        Context context = new Context(0, diagnosticTests, new long[]{0}, "");
         context = new CodeComputer().execute(context);
         Assert.assertEquals(Integer.valueOf(context.getOutput()), Integer.valueOf(0));
     }
@@ -17,7 +17,7 @@ public class CodeComputerTest {
     @Test
     public void test1b() {
         int[] diagnosticTests = new int[]{3,12,6,12,15,1,13,14,13,4,13,99,-1,0,1,9};
-        Context context = new Context(0, diagnosticTests, new int[]{1}, "");
+        Context context = new Context(0, diagnosticTests, new long[]{1}, "");
         context = new CodeComputer().execute(context);
         Assert.assertEquals(Integer.valueOf(context.getOutput()), Integer.valueOf(1));
     }
@@ -25,7 +25,7 @@ public class CodeComputerTest {
     @Test
     public void test2a() {
         int[] diagnosticTests = new int[]{3,3,1105,-1,9,1101,0,0,12,4,12,99,1};
-        Context context = new Context(0, diagnosticTests, new int[]{0}, "");
+        Context context = new Context(0, diagnosticTests, new long[]{0}, "");
         context = new CodeComputer().execute(context);
         Assert.assertEquals(Integer.valueOf(context.getOutput()), Integer.valueOf(0));
     }
@@ -33,7 +33,7 @@ public class CodeComputerTest {
     @Test
     public void test2b() {
         int[] diagnosticTests = new int[]{3,3,1105,-1,9,1101,0,0,12,4,12,99,1};
-        Context context = new Context(0, diagnosticTests, new int[]{1}, "");
+        Context context = new Context(0, diagnosticTests, new long[]{1}, "");
         context = new CodeComputer().execute(context);
         Assert.assertEquals(Integer.valueOf(context.getOutput()), Integer.valueOf(1));
     }
